update to pytorch 0.4
