+++ utils.py
@@ -33,7 +33,7 @@ class strLabelConverter(object):
             length = [len(s) for s in text]
             text = ''.join(text)
             text, _ = self.encode(text)
-        return (torch.IntTensor(text), torch.IntTensor(length))
+        return (text,length)
 
     def decode(self, t, length, raw=False):
         """Decode encoded texts back into strs.
